[ALPS06476902] GPU: secure patch GPUCORE-30405
[PATCH] GPUCORE-30405 Prevent use-after-free after region
tracker out-of-memory

Convert kbase_remove_va_region() to not be able to fail,
by accepting that if we cannot reallocate a free region
in the gap left afterremoving a region, then there is
nothing we can do and should just leavethe gap.

Instead, completely unlink the removed region regardless
so that it cannot be in use in case the caller kfree()s it.
An allocation attempted into that gap (SAME_VA allocations only)
will fail gracefully.

This fixes errata GPUSWERRATA-1207(Bifrost) and GPUSWERRATA-1216(Valhall)

MTK-Commit-Id: 97b173c208cebd070f58f23a9e33c63489c54874

Signed-off-by: Qoo Lin <qoo.lin@mediatek.com>
CR-Id: ALPS06476902
Feature: GPU ARM(gpu_arm)
Change-Id: Ic07cc916bd127f9620e36dee55f957607a041916